requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
Pillow==8.3.2
requests==2.26.0
torch==1.9.1+cpu
torchvision==0.10.1+cpu
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
zipp==3.5.0
